+++ go/action_kit_commons/network/netfault/blackhole.go
@@ -83,10 +83,6 @@ func (o *BlackholeOpts) ipCommands(family family, mode mode) ([]string, error) {
 	return cmds, nil
 }
 
-func (o *BlackholeOpts) tcCommands(_ mode) ([]string, error) {
-	return nil, nil
-}
-
 func (o *BlackholeOpts) String() string {
 	var sb strings.Builder
 	sb.WriteString("blocking traffic ")