+++ go/action_kit_commons/network/netfault/runner_runc_test.go
@@ -48,7 +48,7 @@ func Test_generateAndRunCommands_using_runc_should_serialize(t *testing.T) {
 
 			runner := NewRuncRunner(runcMock, sidecar)
 
-			_ = Apply(context.Background(), runner, &blackholeOpts)
+			_, _ = Apply(context.Background(), runner, &blackholeOpts)
 			defer func(ctx context.Context, runner CommandRunner, opts Opts) {
 				_ = Revert(ctx, runner, opts)
 			}(context.Background(), runner, &blackholeOpts)